libs/util/db/mongoose/src/db_config.py
from os import path

from dotenv import dotenv_values

env_path = '.env'

if not path.exists(env_path):
    raise Exception('.env request_file not found')

config = dotenv_values(env_path)

DATABASE_URL = config['DATABASE_URL']
DATABASE_NAME = config['DATABASE_NAME']
