requirements.txt
pymongo~=4.5.0
passlib~=1.7.4
python-dotenv~=1.0.0
uvicorn~=0.25.0
fastapi~=0.109.0
pydantic~=2.5.3
python-jose
pydantic[email]
